move non-crud actions to actions file
