+++ acslib/ccure/__init__.py
@@ -2,6 +2,7 @@ from typing import Optional
 
 from acslib.base import ACSRequestResponse
 from acslib.base.connection import ACSRequestData, ACSRequestMethod
+from acslib.ccure.actions import CcureAction
 from acslib.ccure.base import CcureACS
 from acslib.ccure.connection import CcureConnection
 from acslib.ccure.crud import (
@@ -19,4 +20,5 @@ class CcureAPI:
         self.clearance = CcureClearance(connection)
         self.credential = CcureCredential(connection)
         self.clearance_item = CcureClearanceItem(connection)
+        self.action = CcureAction(connection)
         self.ccure_object = CcureACS(connection)
